+++ src/androidTest/kotlin/us/frollo/frollosdk/authentication/OAuth2HelperTest.kt
@@ -94,7 +94,7 @@ class OAuth2HelperTest {
     fun testGetAuthorizationRequest() {
         val request = oAuth.getAuthorizationRequest(scopes = listOf("offline_access", "openid", "email"))
         assertNotNull(request)
-        assertEquals(oAuth.oAuth2.clientId, request.clientId)
+        assertEquals(oAuth.config.clientId, request.clientId)
         assertEquals(oAuth.oAuth2.redirectUrl, request.redirectUri.toString())
         assertTrue(request.scopeSet?.containsAll(setOf(OFFLINE_ACCESS, EMAIL, OPENID)) == true)
         assertEquals(oAuth.oAuth2.authorizationUri, request.configuration.authorizationEndpoint)