+++ src/main/kotlin/us/frollo/frollosdk/FrolloSDK.kt
@@ -259,7 +259,7 @@ object FrolloSDK : AuthenticationCallback {
             _goals = Goals(network, database, authentication)
 
             // 16. Setup User Management
-            _userManagement = UserManagement(deviceInfo, network, database, preferences, authentication, authenticationCallback = this)
+            _userManagement = UserManagement(deviceInfo, network, configuration.clientId, database, preferences, authentication, authenticationCallback = this)
 
             // 17. Setup Notifications
             _notifications = Notifications(userManagement, events, messages)