+++ src/main/kotlin/us/frollo/frollosdk/model/api/user/UserResetPasswordRequest.kt
@@ -19,5 +19,6 @@ package us.frollo.frollosdk.model.api.user
 import com.google.gson.annotations.SerializedName
 
 internal data class UserResetPasswordRequest(
-    @SerializedName("email") val email: String
+    @SerializedName("email") val email: String,
+    @SerializedName("client_id") val clientId: String
 )